Refactor active tab handling
